Redesign:  Fix deleted wallet after reset and app restart (#291)
* Add Unified activities entry animation

* Improve wallet wipe functionality

* Minor UI improvement

+++ wallet/src/de/schildbach/wallet/WalletApplication.java
@@ -34,6 +34,7 @@ import android.content.pm.PackageManager.NameNotFoundException;
 import android.media.AudioAttributes;
 import android.net.Uri;
 import android.os.Build;
+import android.os.Handler;
 import android.os.StrictMode;
 import android.preference.PreferenceManager;
 import android.text.format.DateUtils;
@@ -162,11 +163,6 @@ public class WalletApplication extends MultiDexApplication {
         });
         walletFile = getFileStreamPath(Constants.Files.WALLET_FILENAME_PROTOBUF);
         if (walletFileExists()) {
-            if (config.getEnsureWipe()) {
-                //noinspection ResultOfMethodCallIgnored
-                walletFile.delete();
-                config.clear(false);
-            }
             fullInitialization();
         }
         registerDeviceInteractiveReceiver();
@@ -697,26 +693,32 @@ public class WalletApplication extends MultiDexApplication {
     /**
      * Removes all the data and restarts the app showing onboarding screen.
      */
-    @SuppressWarnings("ResultOfMethodCallIgnored")
-    public void wipe(Context context) {
+    public void triggerWipe(final Context context) {
         log.info("Removing all the data and restarting the app.");
 
-        resetBlockchain();
-        wallet.shutdownAutosaveAndWait();
-        stopBlockchainService();
+        startService(new Intent(BlockchainService.ACTION_WIPE_WALLET, null, this, BlockchainServiceImpl.class));
+    }
 
-        walletFile.delete();
+    @SuppressWarnings("ResultOfMethodCallIgnored")
+    public void finalizeWipe() {
+        if (walletFile.exists()) {
+            wallet.shutdownAutosaveAndWait();
+            walletFile.delete();
+        }
+        System.out.println("walletFile.exists(): " + walletFile.exists());
+        if (walletFile.exists()) {
+            walletFile.delete();
+        }
         cleanupFiles();
-        config.clear(true);
+        config.clear();
         PinRetryController.getInstance().clearPinFailPrefs();
 
         File walletBackupFile = getFileStreamPath(Constants.Files.WALLET_KEY_BACKUP_PROTOBUF);
-        if (walletBackupFile.exists())
+        if (walletBackupFile.exists()) {
             walletBackupFile.delete();
-
-        ProcessPhoenix.triggerRebirth(context);
+        }
+        ProcessPhoenix.triggerRebirth(this);
     }
-
 
     public boolean isBackupDisclaimerDismissed() {
         return backupDisclaimerDismissed;
